Add LuaScriptInstance struct (#5)
Now script instances are a struct with a table for storing data. This
makes it possible to always call setter functions on __newindex,
regardless of if there already was a value in the table.
The methods `rawget` and `rawset` were added to index the table
directly, without calling getters/setters or indexing the owner Object.